refactor(ai-gateway): extract partner body reader
